MA-19591 [Android]Add workaround for i.MX8MQ ERR11171
i.MX8MQ is missing the wake_request signals from GIC to GPCv2. This indirectly
breaks cpuidle and suspend/resume support due to inability to wake target cores on IPIs.

Here is the link to the errata (see e11171):

https://www.nxp.com/docs/en/errata/IMX8MDQLQ_0N14W.pdf

Because some of these changes cannot be uploaded to linux, i.MX8MQ will not
be able to support cpuidle and suspend/resume after we enable GKI.

I changed the status of power from mem to freeze. The predictable result
is that the freeze state consumes more power than the mem state.

Change-Id: I947d381de653eefdbdb82c2c35c832504ccdb698
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>